+++ src/effect_send.rs
@@ -1,6 +1,5 @@
 use std::rc::Rc;
 
-use super::effect::{EffectProcessIter, EffectRenderState};
 use super::effect_node::EffectNode;
 use super::partial::Partial;
